# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "MedCenter Back",
    "description": "API for MedCenter",
    "version": "1.0.0"
  },
  "host": "localhost:8088",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/pacients/{id}/history": {
      "get": {
        "tags": ["Pacients"],
        "summary": "Получение медицинской истории пациента",
        "description": "Список записей медицинской истории (диагнозы, назначения) по id пациента (карты пациента). Сортировка по дате, новые первые.",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID карты пациента (pacients.id)"
          }
        ],
        "responses": {
          "200": {
            "description": "Список записей",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/MedicalHistoryRecord" }
            }
          },
          "400": { "description": "Некорректный id", "schema": { "$ref": "#/definitions/Message" } },
          "401": { "description": "Требуется авторизация", "schema": { "$ref": "#/definitions/Message" } }
        }
      },
      "post": {
        "tags": ["Pacients"],
        "summary": "Создание записи в медицинской истории",
        "description": "Добавление новой записи (диагноз, назначение) для пациента. Требуется авторизация.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID карты пациента (pacients.id)"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/CreateMedicalHistory" }
          }
        ],
        "responses": {
          "201": {
            "description": "Запись создана",
            "schema": { "$ref": "#/definitions/MedicalHistoryRecord" }
          },
          "400": {
            "description": "Ошибка валидации или неверный id пациента/доктора",
            "schema": { "$ref": "#/definitions/Message" }
          },
          "401": { "description": "Требуется авторизация", "schema": { "$ref": "#/definitions/Message" } }
        }
      }
    },
...
    "/create": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Создания анализа",
        "description": "Создание анализа для пациента",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Новость была создана",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "400": {
            "description": "Не удалось создать новость",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/list": {
      "post": {
        "tags": [
          "News"
        ],
        "summary": "Список новостей",
        "description": "Получения списка новостей по фильтрам",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Фильтры для получения новостей",
            "required": false,
            "schema": {
              "$ref": "#/definitions/NewsFilters"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/NewsList"
            },
            "description": "OK"
          },
          "400": {
            "schema": {
              "$ref": "#/definitions/NewsList"
            },
            "description": "Bad Request"
          }
        }
      },
      "get": {
        "tags": [
          "Analysis"
        ],
        "summary": "Получение списка анализов",
        "description": "Получение списка всех анализов с количеством",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Фильтры для получения анализов",
            "required": false,
            "schema": {
              "$ref": "#/definitions/AnalysFilters"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/AnalysisList"
            },
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/update": {
      "put": {
        "tags": [
          "News"
        ],
        "summary": "Обновление информации по новости",
        "description": "Частичное или полное обновление информации по новосте",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Поля под обновление",
            "required": false,
            "schema": {
              "$ref": "#/definitions/NewsFilters"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о новосте была обновлена",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "400": {
            "description": "Не удалось обновить информацию по новости",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        }
      }
    },
    "/delete": {
      "delete": {
        "tags": [
          "News"
        ],
        "summary": "Удаление новости",
        "description": "Удаление новости администратором",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Идентификатор новости",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Новость была удалена",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "400": {
            "description": "Не удалось удалить новость",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Message": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Сообщение"
        }
      }
    },
...
    "MedicalHistoryRecord": {
      "type": "object",
      "description": "Запись медицинской истории (диагноз, назначение)",
      "properties": {
        "id": { "type": "integer", "example": 1 },
        "createdAt": { "type": "string", "format": "date-time", "description": "Дата записи" },
        "diseases": { "type": "string", "maxLength": 165, "description": "Диагноз", "example": "ОРВИ" },
        "description": { "type": "string", "description": "Назначение/рекомендации", "example": "Обильное питье, постельный режим" },
        "status": { "type": "string", "enum": ["active", "completed"], "example": "completed" },
        "doctor": {
          "type": "object",
          "description": "Доктор",
          "properties": {
            "id": { "type": "integer" },
            "fullName": { "type": "string", "example": "Доктор Сидоров" },
            "clinicType": { "type": "object", "properties": { "name": { "type": "string", "example": "Терапевт" } } }
          }
        }
      }
    },
    "CreateMedicalHistory": {
      "type": "object",
      "required": ["diseases", "doctorId"],
      "properties": {
        "diseases": { "type": "string", "maxLength": 165, "description": "Диагноз", "example": "ОРВИ" },
        "description": { "type": "string", "description": "Назначение/рекомендации", "example": "Обильное питье" },
        "doctorId": { "type": "integer", "description": "ID доктора" },
        "status": { "type": "string", "enum": ["active", "completed"], "example": "active" }
      }
    },
...
    "NewsList": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string",
                "example": "Заголовок новости"
              },
              "description": {
                "type": "string",
                "example": "Описание новости"
              },
              "type": {
                "default": "GENERAL",
                "type": "string",
                "enum": [
                  "GENERAL",
                  "MED",
                  "SOCIAL",
                  "EVENTS",
                  "CONFERENCE",
                  "PERSONAL"
                ]
              },
              "createdAt": {
                "type": "string",
                "example": "24-12-2025"
              },
              "updatedAt": {
                "type": "string",
                "example": "24-12-2025"
              }
            }
          }
        },
        "total": {
          "type": "number",
          "example": 1
        }
      }
    },
    "NewsFilters": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "Заголовок поиска"
        },
        "description": {
          "type": "string",
          "example": "Описание поиска"
        },
        "type": {
          "default": "GENERAL",
          "type": "string",
          "enum": [
            "GENERAL",
            "MED",
            "SOCIAL",
            "EVENTS",
            "CONFERENCE",
            "PERSONAL"
          ]
        },
        "createdAt": {
          "type": "string",
          "example": "24-12-2025"
        },
        "updatedAt": {
          "type": "string",
          "example": "24-12-2025"
        }
      }
    },
...
    "AnalysisList": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "Тип"
              },
              "text": {
                "type": "string",
                "example": "Текст"
              },
              "status": {
                "type": "string",
                "example": "DEAD"
              },
              "assignedDate": {
                "type": "string",
                "example": "2025-10-10"
              },
              "takenDate": {
                "type": "string",
                "example": "2025-10-10"
              },
              "readyDate": {
                "type": "string",
                "example": "2025-10-10"
              },
              "results": {
                "type": "object",
                "properties": {
                  "blood": {
                    "type": "string",
                    "example": "C+"
                  }
                }
              },
              "costs": {
                "type": "number",
                "example": 200
              },
              "updatedAt": {
                "type": "string",
                "example": "2025-10-11"
              },
              "createdAt": {
                "type": "string",
                "example": "2025-10-11"
              }
            }
          }
        },
        "total": {
          "type": "number",
          "example": 1
        }
      }
    },
    "AnalysFilters": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "кровь"
        },
        "costs": {
          "type": "number",
          "example": 1025
        },
        "status": {
          "type": "string",
          "example": "HEALTHY"
        }
      }
    },
...
  }
}